refactor: extract splitting, sorting and bulk conversion into a mixin
The 23 methods under the "How to split and sort filenames" and "Transformations"
banners move verbatim into src/pdsfile/_sorting.py as _SortingMixin, which the
class statement takes as a base in its alphabetical slot. The bodies are
byte-for-byte what they were, including both in-class banners; nothing in this
commit is a content edit inside a moved body.

Two banner blocks, one module. The parent plan names exactly two module files for
this PR and a third would deviate from the stated deliverable; splitting/sorting
and the twelve bulk-conversion helpers are also one domain, operations over lists
of basenames, logical paths, abspaths and PdsFile objects with no I/O of their
own. The module docstring says so, because the filename alone undersells it.

The sort configuration does not move: SORT_ORDER, SORT_KEY and the
sort_labels_after/sort_dirs_first/sort_dirs_last/sort_info_first setters stay on
PdsFile, where these methods read them off, and are on PR-22's stay-list.

The free-variable sweep names four module-level references: os, _clean_join,
abspath_for_logical_path and logical_path_from_abspath. No class object appears,
so unlike _opus.py this module needs no function-local deferred import. All four
names keep live callers in pdsfile.py and so keep their plain import form there.

The new module's per-file-ignores entry is a split of pdsfile.py's: E701 and
RUF005 are the codes its moved lines carry, both already forgiven for the same
lines in the same file, and every code in pdsfile.py's entry conserves.
