+++ pyproject.toml
@@ -167,6 +167,7 @@ extend-ignore = ["PT011", "SIM105", "SIM108"]
 "src/pdsfile/_opus.py" = ["RUF005", "UP024"]
 "src/pdsfile/_path_utils.py" = ["E701", "F841"]
 "src/pdsfile/_shelves.py" = ["B904", "B905", "F841", "RUF059", "SIM103", "UP024", "UP031"]
+"src/pdsfile/_sorting.py" = ["E701", "RUF005"]
 "src/pdsfile/holdings_maintenance/pds3/crlf.py" = ["PT028"]
 "src/pdsfile/holdings_maintenance/pds3/pdsarchives.py" = ["A001", "C405", "F841", "I001", "RUF059", "SIM115", "UP031"]
 "src/pdsfile/holdings_maintenance/pds3/pdschecksums.py" = ["A001", "B006", "B007", "B012", "C405", "I001", "N806", "RUF059", "SIM115", "UP015", "UP024", "UP031"]
